+++ src/local_path_planner.cpp
@@ -27,18 +27,18 @@ void Robot::update_pose(const geometry_msgs::PoseStamped pose)
 }
 
 // 直前の制御入力を記録
-void Robot::set_speed(const float velocity, const float yawrate)
+void Robot::set_speed(const double velocity, const double yawrate)
 {
     state_.velocity = velocity;
     state_.yawrate  = yawrate;
 }
 
 // メンバ関数の値を返却する関数
-float Robot::x()        { return state_.x; }
-float Robot::y()        { return state_.y; }
-float Robot::yaw()      { return state_.yaw; }
-float Robot::velocity() { return state_.velocity; }
-float Robot::yawrate()  { return state_.yawrate; }
+double Robot::x()        { return state_.x; }
+double Robot::y()        { return state_.y; }
+double Robot::yaw()      { return state_.yaw; }
+double Robot::velocity() { return state_.velocity; }
+double Robot::yawrate()  { return state_.yawrate; }
 
 
 // ===== クラスLocalPathPlanner =====
@@ -99,7 +99,7 @@ void LocalPathPlanner::ob_poses_callback(const geometry_msgs::PoseArray::ConstPt
 // }
 
 // Roombaの制御入力を行う
-void LocalPathPlanner::roomba_control(const float velocity, const float yawrate)
+void LocalPathPlanner::roomba_control(const double velocity, const double yawrate)
 {
     cmd_speed_.mode           = 11; // 任意の動きを実行するモード
     cmd_speed_.cntl.linear.x  = velocity;
@@ -109,7 +109,7 @@ void LocalPathPlanner::roomba_control(const float velocity, const float yawrate)
 }
 
 // 予測軌跡作成時における仮想ロボットを移動
-void LocalPathPlanner::move(State& state, const float velocity, const float yawrate)
+void LocalPathPlanner::move(State& state, const double velocity, const double yawrate)
 {
     state.yaw      += yawrate * dt_;
     state.yaw       = optimize_angle(state.yaw);
@@ -120,7 +120,7 @@ void LocalPathPlanner::move(State& state, const float velocity, const float yawr
 }
 
 // 適切な角度(-M_PI ~ M_PI)を返す
-float LocalPathPlanner::optimize_angle(float angle)
+double LocalPathPlanner::optimize_angle(double angle)
 {
     if(M_PI  < angle) angle -= 2.0*M_PI;
     if(angle < -M_PI) angle += 2.0*M_PI;
@@ -132,10 +132,10 @@ float LocalPathPlanner::optimize_angle(float angle)
 void LocalPathPlanner::calc_dynamic_window()
 {
     // 車両モデルによるWindow
-    float Vs[] = {min_vel_, max_vel_, -max_yawrate_, max_yawrate_};
+    double Vs[] = {min_vel_, max_vel_, -max_yawrate_, max_yawrate_};
 
     // 運動モデルによるWindow
-    float Vd[] = {roomba_.velocity() - max_accel_*dt_,
+    double Vd[] = {roomba_.velocity() - max_accel_*dt_,
                   roomba_.velocity() + max_accel_*dt_,
                   roomba_.yawrate()  - max_dyawrate_*dt_,
                   roomba_.yawrate()  - max_dyawrate_*dt_};
@@ -148,7 +148,7 @@ void LocalPathPlanner::calc_dynamic_window()
 }
 
 // 予測軌跡を作成
-std::vector<State> LocalPathPlanner::calc_trajectory(const float velocity, const float yawrate)
+std::vector<State> LocalPathPlanner::calc_trajectory(const double velocity, const double yawrate)
 {
     std::vector<State> trajectory;           // 軌跡格納用の動的配列
     State state = {0.0, 0.0, 0.0, 0.0, 0.0}; // 軌跡作成用の仮想ロボット
@@ -159,7 +159,7 @@ std::vector<State> LocalPathPlanner::calc_trajectory(const float velocity, const
     // state.yaw = roomba_.yaw();
 
     // 軌跡を格納
-    for(float t=0.0; t<=predict_time_; t+=dt_)
+    for(double t=0.0; t<=predict_time_; t+=dt_)
     {
         move(state, velocity, yawrate);
         trajectory.push_back(state);
@@ -169,19 +169,19 @@ std::vector<State> LocalPathPlanner::calc_trajectory(const float velocity, const
 }
 
 // 最適な制御入力を計算
-std::vector<float> LocalPathPlanner::calc_final_input()
+std::vector<double> LocalPathPlanner::calc_final_input()
 {
-    std::vector<float> input{0.0, 0.0}; // {velocity, yawrate}
-    float max_score = 0.0;              // 評価値の最大値格納用
+    std::vector<double> input{0.0, 0.0}; // {velocity, yawrate}
+    double max_score = 0.0;              // 評価値の最大値格納用
     calc_dynamic_window();              // ダイナミックウィンドウを計算
 
     // 並進速度と旋回速度のすべての組み合わせを評価
-    for(float velocity=dw_.min_vel; velocity<=dw_.max_vel; velocity+=vel_reso_)
+    for(double velocity=dw_.min_vel; velocity<=dw_.max_vel; velocity+=vel_reso_)
     {
-        for(float yawrate=dw_.min_yawrate; yawrate<=dw_.max_yawrate; yawrate+=yawrate_reso_)
+        for(double yawrate=dw_.min_yawrate; yawrate<=dw_.max_yawrate; yawrate+=yawrate_reso_)
         {
             const std::vector<State> trajectory = calc_trajectory(velocity, yawrate); // 予測軌跡の生成
-            const float score = calc_evaluation(trajectory); // 予測軌跡に対する評価値の計算
+            const double score = calc_evaluation(trajectory); // 予測軌跡に対する評価値の計算
 
             // 最大値の更新
             if(max_score < score)
@@ -199,43 +199,43 @@ std::vector<float> LocalPathPlanner::calc_final_input()
 }
 
 // 評価関数を計算
-float LocalPathPlanner::calc_evaluation(const std::vector<State> traj)
+double LocalPathPlanner::calc_evaluation(const std::vector<State> traj)
 {
-    const float heading_score  = weight_heading_ * calc_heading_eval(traj);
-    const float distance_score = weight_dist_    * calc_dist_eval(traj);
-    const float velocity_score = weight_vel_     * calc_vel_eval(traj);
+    const double heading_score  = weight_heading_ * calc_heading_eval(traj);
+    const double distance_score = weight_dist_    * calc_dist_eval(traj);
+    const double velocity_score = weight_vel_     * calc_vel_eval(traj);
 
-    const float total_score    = heading_score + distance_score + velocity_score;
+    const double total_score    = heading_score + distance_score + velocity_score;
 
     return total_score;
 }
 
 // headingの評価関数を計算
-float LocalPathPlanner::calc_heading_eval(const std::vector<State> traj)
+double LocalPathPlanner::calc_heading_eval(const std::vector<State> traj)
 {
     // 最終時刻のロボットの方位
-    const float theta = traj.back().yaw;
+    const double theta = traj.back().yaw;
 
     // 最終時刻の位置に対するゴールの方位
-    const float goal_theta = atan2(local_goal_.point.y - traj.back().y, local_goal_.point.x - traj.back().x);
+    const double goal_theta = atan2(local_goal_.point.y - traj.back().y, local_goal_.point.x - traj.back().x);
 
     // ゴールまでの方位差分
-    float target_theta = 0.0;
+    double target_theta = 0.0;
     if(goal_theta > theta)
         target_theta = goal_theta - theta;
     else
         target_theta = theta - goal_theta;
 
     // headingの評価値
-    const float heading_eval = (M_PI - abs(optimize_angle(target_theta)))/M_PI; // 正規化
+    const double heading_eval = (M_PI - abs(optimize_angle(target_theta)))/M_PI; // 正規化
 
     return heading_eval;
 }
 
 // distの評価関数を計算
-float LocalPathPlanner::calc_dist_eval(const std::vector<State> traj)
+double LocalPathPlanner::calc_dist_eval(const std::vector<State> traj)
 {
-    float min_dist = search_range_; // 最も近い障害物との距離
+    double min_dist = search_range_; // 最も近い障害物との距離
 
     // pathの点と障害物のすべての組み合わせを探索
     for(auto& state : traj)
@@ -243,9 +243,9 @@ float LocalPathPlanner::calc_dist_eval(const std::vector<State> traj)
         for(auto& ob_pose : ob_poses_.poses)
         {
             // pathのうちの１点と障害物の距離を計算
-            const float dx   = ob_pose.position.x - state.x;
-            const float dy   = ob_pose.position.y - state.y;
-            const float dist = sqrt(powf(dx, 2.0)+powf(dy, 2.0));
+            const double dx   = ob_pose.position.x - state.x;
+            const double dy   = ob_pose.position.y - state.y;
+            const double dist = sqrt(powf(dx, 2.0)+powf(dy, 2.0));
 
             // 最小値の更新
             if(dist < min_dist)
@@ -257,7 +257,7 @@ float LocalPathPlanner::calc_dist_eval(const std::vector<State> traj)
 }
 
 // velocityの評価関数を計算
-float LocalPathPlanner::calc_vel_eval(const std::vector<State> traj)
+double LocalPathPlanner::calc_vel_eval(const std::vector<State> traj)
 {
     if(0 < traj.back().velocity) // 前進
         return traj.back().velocity/max_vel_; // 正規化
@@ -268,9 +268,9 @@ float LocalPathPlanner::calc_vel_eval(const std::vector<State> traj)
 // ゴールに着くまでTrueを返す
 bool LocalPathPlanner::can_move()
 {
-    const float dx = local_goal_.point.x - roomba_.x();
-    const float dy = local_goal_.point.y - roomba_.y();
-    const float dist_to_goal = sqrt(powf(dx, 2.0) + powf(dy, 2.0)); // 現在位置からゴールまでの距離
+    const double dx = local_goal_.point.x - roomba_.x();
+    const double dy = local_goal_.point.y - roomba_.y();
+    const double dist_to_goal = sqrt(powf(dx, 2.0) + powf(dy, 2.0)); // 現在位置からゴールまでの距離
 
     if(dist_to_goal > goal_tolerance_)
         return true;
@@ -286,7 +286,7 @@ void LocalPathPlanner::process()
     {
         if(can_move())
         {
-            const std::vector<float> input = calc_final_input();
+            const std::vector<double> input = calc_final_input();
             roomba_control(input[0], input[1]);
         }
         else
